fix:update grinningScore every 10 seconds and when paused
