+++ src/contexts/coreFunctionsContext.tsx
@@ -1,12 +1,9 @@
 import React, { createContext, useState, FC } from 'react';
 import { MoviePlayerState } from 'types';
-import { Movie } from 'types/dataTypes';
 
 type CoreFunctionsState = {
   moviePlayerState: MoviePlayerState;
   setMoviePlayerState: (input: MoviePlayerState) => void;
-  grinningScore: number;
-  setGrinningScore: (input: number) => void;
 };
 
 export const CoreFunctionsContext = createContext({} as CoreFunctionsState);
@@ -17,15 +14,12 @@ const CoreFucntionsProvider: FC = ({ children }) => {
     currentTime: 0,
     duration: 0,
   });
-  const [grinningScore, setGrinningScore] = useState<number>(0);
 
   return (
     <CoreFunctionsContext.Provider
       value={{
         moviePlayerState: moviePlayerState,
         setMoviePlayerState: setMoviePlayerState,
-        grinningScore: grinningScore,
-        setGrinningScore: setGrinningScore,
       }}
     >
       {children}
